+++ tools/test-js-with-node
@@ -127,6 +127,7 @@ EXEMPT_FILES = make_set(
         "static/js/pm_list_dom.js",
         "static/js/poll_widget.js",
         "static/js/popover_menus.js",
+        "static/js/popover_menus_data.js",
         "static/js/popovers.js",
         "static/js/read_receipts.js",
         "static/js/ready.ts",
